requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.1.0
attrs==23.1.0
Authlib==1.2.1
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
distro==1.8.0
distyll-info==0.0.1
exceptiongroup==1.2.0
gitdb==4.0.11
GitPython==3.1.40
grpcio==1.60.0
grpcio-tools==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mutagen==1.47.0
numpy==1.26.2
openai==1.3.8
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==4.25.1
pyarrow==14.0.1
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pypdf==3.17.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
streamlit==1.29.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
weaviate-client==4.4b1
websockets==12.0
yt-dlp==2023.11.16
zipp==3.17.0
